refactor(server): split the device flow into focused per-flow files
deviceflowhandlers.go mixed the device-authorization page/request, the browser
callback, and the device_code grant in one file — outside the grant_*.go family
and without the per-flow header comments the rest of the package now has.

Split it into device_authorize.go (the /device page, /device/code request, and
user-code verification, which share the device template), device_callback.go
(the /device/callback token-minting callback), and grant_devicecode.go (the
device_code grant, a token.go dispatch target alongside the other grants). Tests
are split to match. Pure code movement; the same 176 tests pass.
